requirements.txt
numpy
networkx
python-igraph
matplotlib
torch
tqdm